add sod pvp sets, hide more old sets
