--- sim/hunter/item_sets.go
@@ -1,55 +1,0 @@
-package hunter
-
-import (
-	"github.com/wowsims/sod/sim/core"
-	"github.com/wowsims/sod/sim/core/stats"
-)
-
-///////////////////////////////////////////////////////////////////////////
-//                            SoD Phase 3 Item Sets
-///////////////////////////////////////////////////////////////////////////
-
-var ItemSetDreadHuntersChain = core.NewItemSet(core.ItemSet{
-	Name: "Dread Hunter's Chain",
-	Bonuses: map[int32]core.ApplyEffect{
-		2: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.AttackPower, 20)
-			c.AddStat(stats.RangedAttackPower, 20)
-		},
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddBonusRangedCritRating(2)
-		},
-	},
-})
-
-var ItemSetBloodGuardsChain = core.NewItemSet(core.ItemSet{
-	Name: "Blood Guard's Chain",
-	Bonuses: map[int32]core.ApplyEffect{
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.Stamina, 15)
-		},
-		6: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.AttackPower, 30)
-			c.AddStat(stats.RangedAttackPower, 20)
-		},
-	},
-})
-
-var ItemSetKnightLieutenantsChain = core.NewItemSet(core.ItemSet{
-	Name: "Knight-Lieutenant's Chain",
-	Bonuses: map[int32]core.ApplyEffect{
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.Stamina, 15)
-		},
-		6: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.AttackPower, 30)
-			c.AddStat(stats.RangedAttackPower, 20)
-		},
-	},
-})
