+++ sim/hunter/item_sets_pve.go
@@ -1,0 +1,29 @@
+package hunter
+
+import (
+	"github.com/wowsims/sod/sim/core"
+	"github.com/wowsims/sod/sim/core/stats"
+)
+
+///////////////////////////////////////////////////////////////////////////
+//                            SoD Phase 3 Item Sets
+///////////////////////////////////////////////////////////////////////////
+
+var ItemSetDreadHuntersChain = core.NewItemSet(core.ItemSet{
+	Name: "Dread Hunter's Chain",
+	Bonuses: map[int32]core.ApplyEffect{
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.AttackPower, 20)
+			c.AddStat(stats.RangedAttackPower, 20)
+		},
+		3: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddBonusRangedCritRating(2)
+		},
+	},
+})
+
+///////////////////////////////////////////////////////////////////////////
+//                            SoD Phase 4 Item Sets
+///////////////////////////////////////////////////////////////////////////
