+++ sim/mage/item_sets_pvp.go
@@ -1,0 +1,50 @@
+package mage
+
+import (
+	"github.com/wowsims/sod/sim/core"
+	"github.com/wowsims/sod/sim/core/stats"
+)
+
+///////////////////////////////////////////////////////////////////////////
+//                            SoD Phase 4 Item Sets
+///////////////////////////////////////////////////////////////////////////
+
+var ItemSetChampionsRegalia = core.NewItemSet(core.ItemSet{
+	Name: "Champion's Regalia",
+	Bonuses: map[int32]core.ApplyEffect{
+		// Increases damage and healing done by magical spells and effects by up to 23.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.SpellPower, 23)
+		},
+		// Reduces the cooldown of your Blink spell by 1.5 sec.
+		4: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// +20 Stamina.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 20)
+		},
+	},
+})
+
+var ItemSetLieutenantCommandersRegalia = core.NewItemSet(core.ItemSet{
+	Name: "Lieutenant Commander's Regalia",
+	Bonuses: map[int32]core.ApplyEffect{
+		// Increases damage and healing done by magical spells and effects by up to 23.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.SpellPower, 23)
+		},
+		// Reduces the cooldown of your Blink spell by 1.5 sec.
+		4: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// +20 Stamina.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 20)
+		},
+	},
+})
