+++ sim/paladin/item_sets_pve.go
@@ -1,0 +1,42 @@
+package paladin
+
+import (
+	"github.com/wowsims/sod/sim/core"
+	"github.com/wowsims/sod/sim/core/stats"
+)
+
+///////////////////////////////////////////////////////////////////////////
+//                            SoD Phase 3 Item Sets
+///////////////////////////////////////////////////////////////////////////
+
+var ItemSetObsessedProphetsPlate = core.NewItemSet(core.ItemSet{
+	Name: "Obsessed Prophet's Plate",
+	Bonuses: map[int32]core.ApplyEffect{
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.MeleeCrit, 1*core.CritRatingPerCritChance)
+			c.AddStat(stats.SpellCrit, 1*core.SpellCritRatingPerCritChance)
+		},
+		3: func(agent core.Agent) {
+			// handled via paladin.holyCrit()
+		},
+	},
+})
+
+var _ = core.NewItemSet(core.ItemSet{
+	Name: "Emerald Encrusted Battleplate",
+	Bonuses: map[int32]core.ApplyEffect{
+		3: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 10)
+		},
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.HealingPower, 22)
+		},
+	},
+})
+
+///////////////////////////////////////////////////////////////////////////
+//                            SoD Phase 4 Item Sets
+///////////////////////////////////////////////////////////////////////////
