+++ sim/priest/item_sets_pvp.go
@@ -1,0 +1,90 @@
+package priest
+
+import (
+	"github.com/wowsims/sod/sim/core"
+	"github.com/wowsims/sod/sim/core/stats"
+)
+
+///////////////////////////////////////////////////////////////////////////
+//                            SoD Phase 4 Item Sets
+///////////////////////////////////////////////////////////////////////////
+
+var ItemSetChampionsRaiment = core.NewItemSet(core.ItemSet{
+	Name: "Champion's Raiment",
+	Bonuses: map[int32]core.ApplyEffect{
+		// Increases damage and healing done by magical spells and effects by up to 23.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.SpellPower, 23)
+		},
+		// Increases the duration of your Psychic Scream spell by 1 sec.
+		4: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// +20 Stamina.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 20)
+		},
+	},
+})
+
+var ItemSetChampionsInvestiture = core.NewItemSet(core.ItemSet{
+	Name: "Champion's Investiture",
+	Bonuses: map[int32]core.ApplyEffect{
+		// Increases healing done by spells and effects by up to 44.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.HealingPower, 44)
+		},
+		// Increases the duration of your Psychic Scream spell by 1 sec.
+		4: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// +20 Stamina.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 20)
+		},
+	},
+})
+
+var ItemSetLieutenantCommandersRaiment = core.NewItemSet(core.ItemSet{
+	Name: "Lieutenant Commander's Raiment",
+	Bonuses: map[int32]core.ApplyEffect{
+		// Increases damage and healing done by magical spells and effects by up to 23.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.SpellPower, 23)
+		},
+		// Increases the duration of your Psychic Scream spell by 1 sec.
+		4: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// +20 Stamina.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 20)
+		},
+	},
+})
+
+var ItemSetLieutenantCommandersInvestiture = core.NewItemSet(core.ItemSet{
+	Name: "Lieutenant Commander's Investiture",
+	Bonuses: map[int32]core.ApplyEffect{
+		// Increases healing done by spells and effects by up to 44.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.HealingPower, 44)
+		},
+		// Increases the duration of your Psychic Scream spell by 1 sec.
+		4: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// +20 Stamina.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 20)
+		},
+	},
+})
