+++ sim/warrior/item_sets_pve.go
@@ -1,0 +1,5 @@
+package warrior
+
+///////////////////////////////////////////////////////////////////////////
+//                            SoD Phase 4 Item Sets
+///////////////////////////////////////////////////////////////////////////
