+++ sim/warrior/item_sets_pvp.go
@@ -1,0 +1,56 @@
+package warrior
+
+import (
+	"github.com/wowsims/sod/sim/core"
+	"github.com/wowsims/sod/sim/core/stats"
+)
+
+///////////////////////////////////////////////////////////////////////////
+//                            SoD Phase 4 Item Sets
+///////////////////////////////////////////////////////////////////////////
+
+var ItemSetChampionsBattlegear = core.NewItemSet(core.ItemSet{
+	Name: "Champion's Battlegear",
+	Bonuses: map[int32]core.ApplyEffect{
+		// +40 Attack Power.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStats(stats.Stats{
+				stats.AttackPower:       40,
+				stats.RangedAttackPower: 40,
+			})
+		},
+		// Reduces the cooldown of your Intercept ability by 5 sec.
+		4: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// +20 Stamina.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 20)
+		},
+	},
+})
+
+var ItemSetLieutenantCommandersBattlegear = core.NewItemSet(core.ItemSet{
+	Name: "Lieutenant Commander's Battlegear",
+	Bonuses: map[int32]core.ApplyEffect{
+		// +40 Attack Power.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStats(stats.Stats{
+				stats.AttackPower:       40,
+				stats.RangedAttackPower: 40,
+			})
+		},
+		// Reduces the cooldown of your Intercept ability by 5 sec.
+		4: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// +20 Stamina.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Stamina, 20)
+		},
+	},
+})
